requirements.txt
appdirs==1.4.4
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bounded-pool-executor==0.0.3
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.1.0
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
filelock==3.12.2
fonttools==4.40.0
gdown==4.7.1
idna==3.4
inflate64==0.3.1
ipykernel==6.23.2
ipython==8.14.0
jedi==0.18.2
joblib==1.2.0
jsonlines==3.1.0
jupyter_client==8.2.0
jupyter_core==5.3.1
kiwisolver==1.4.4
lxml==4.9.2
markdown-it-py==3.0.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multivolumefile==0.2.3
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.2
parso==0.8.3
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.3
polars==0.18.2
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.20.5
pyarrow==12.0.1
pybcj==1.0.1
pycryptodomex==3.18.0
pydantic==1.10.9
Pygments==2.15.1
pyparsing==3.0.9
pyppmd==1.0.0
pyshark==0.6
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.1.0
pyzstd==0.15.8
requests==2.31.0
rich==13.4.2
seaborn==0.12.2
six==1.16.0
soupsieve==2.4.1
stack-data==0.6.2
termcolor==2.3.0
texttable==1.6.7
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
